frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]


[tool.pyright]
venvPath = "."
venv = ".venv"
